Add files via upload
Uploaded simulation study results where the target species (typically) reach the hooks ahead of competing species. Effectively, this results in the target species being 'immune' to hook competition, but not to gear saturation and so we can assess the performance of the censored Poisson estimators against a simple CPUE-based estimator which we expect to perform well.